requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-crispy-forms==2.0
crispy-bootstrap5==0.7
Pillow==10.1.0
python-decouple==3.8
django-allauth==0.57.0 